requirements.txt
# Server specific dependencies
fastapi
uvicorn[standard]
cohere
python-dotenv
lmdb>=1.6.0

# Original dependencies / Other project dependencies
datasets>=3.6.0
numpy>=1.24.0
tqdm>=4.66.0
psutil>=5.9.0

# CUDA-dependent libraries (ensure these match your environment)
cupy-cuda12x>=12.0.0
cuvs-cu12>=25.4.0

